Move old wait built-in implementation to separate module
